+++ src/main/resources/templates/comment.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>댓글 목록</title>
+</head>
+<body>
+<h1>댓글 목록</h1>
+<ul>
+    <li th:each="comment : ${comments}">
+        <p th:text="${comment.content}"></p>
+        <p th:text="${comment.username}"></p>
+        <!-- 이하 필요한 댓글 속성들을 출력 -->
+    </li>
+</ul>
+</body>
+</html>
